+++ src/run.py
@@ -46,9 +46,12 @@ class TemplateContainer:
 
 def main(
     *,
-    organization: str,
-    repository: str,
     name: str,
+    organization: str,
+    project_type: str,
+    repository: str,
+    synchronize_contribution_guide: bool,
+    synchronize_installation_guide: bool,
     synchronize_issue_templates: bool,
     synchronize_pull_request_template: bool,
     synchronize_release_drafter_template: bool,
@@ -67,7 +70,37 @@ def main(
             file_name="bug-report.yml",
             output_dir=Path(".github/ISSUE_TEMPLATE"),
             active=synchronize_issue_templates,
-            arguments={"organization": organization, "repository": repository, "name": name},
+            arguments={"name": name, "organization": organization, "repository": repository},
+        ),
+        TemplateContainer(
+            file_name="CONTRIBUTING.md",
+            output_dir=Path(".github/CONTRIBUTING.md"),
+            template_name="CONTRIBUTING.md",
+            active=synchronize_contribution_guide,
+            arguments={"name": name, "repository": repository},
+        ),
+        TemplateContainer(
+            file_name="CONTRIBUTING.md",
+            output_dir=Path("docs/CONTRIBUTING.md"),
+            template_name="docs_contributing.md",
+            active=synchronize_contribution_guide,
+            arguments={
+                "name": name,
+                "organization": organization,
+                "project_type": project_type,
+                "repository": repository,
+            },
+        ),
+        TemplateContainer(
+            file_name="installation.md",
+            output_dir=Path("docs/installation.md"),
+            active=synchronize_installation_guide,
+            arguments={
+                "name": name,
+                "organization": organization,
+                "project_type": project_type,
+                "repository": repository,
+            },
         ),
         TemplateContainer(
             file_name="feature-request.yml",
@@ -101,10 +134,10 @@ def main(
             arguments={"organization": organization, "repository": repository},
         ),
         TemplateContainer(
-            file_name="support.md",
+            file_name="SUPPORT.md",
             output_dir=Path(".github"),
             active=synchronize_support_resources,
-            arguments={"organization": organization, "repository": repository, "name": name},
+            arguments={"name": name, "organization": organization, "repository": repository},
         ),
     ]
 
@@ -145,10 +178,23 @@ def _convert_to_bool(value: str) -> bool:
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument(
+        "--name",
+        type=str,
+        required=True,
+        help="Stylized name of the package (e.g., 'Core' or 'DDSIM')",
+    )
+    parser.add_argument(
         "--organization",
         type=str,
         required=True,
         help="Name of the repository's organization",
+    )
+    parser.add_argument(
+        "--project_type",
+        type=str,
+        choices=["c++-python", "pure-python"],
+        required=True,
+        help="Type of the project",
     )
     parser.add_argument(
         "--repository",
@@ -157,10 +203,16 @@ if __name__ == "__main__":
         help="Name of the repository",
     )
     parser.add_argument(
-        "--name",
-        type=str,
-        required=True,
-        help="Stylized name of the package (e.g., 'Core' or 'DDSIM')",
+        "--synchronize_contribution_guide",
+        default=True,
+        type=_convert_to_bool,
+        help="Whether to synchronize the contribution guide",
+    )
+    parser.add_argument(
+        "--synchronize_installation_guide",
+        default=True,
+        type=_convert_to_bool,
+        help="Whether to synchronize the installation guide",
     )
     parser.add_argument(
         "--synchronize_issue_templates",
@@ -207,9 +259,12 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     main(
-        organization=args.organization,
-        repository=args.repository,
         name=args.name,
+        organization=args.organization,
+        project_type=args.project_type,
+        repository=args.repository,
+        synchronize_contribution_guide=args.synchronize_contribution_guide,
+        synchronize_installation_guide=args.synchronize_installation_guide,
         synchronize_issue_templates=args.synchronize_issue_templates,
         synchronize_pull_request_template=args.synchronize_pull_request_template,
         synchronize_release_drafter_template=args.synchronize_release_drafter_template,